Save Query in Rotation
The activity didn't save the query when there was a rotation. Because of that, I made some changes: First, I added the onSaveInstanceState method, where I add the query. Then, when the Activity is in the onCreate method, I save the Bundle (if it exists) and then, when the SearchView is created, I set the saved query to the SearchView